Add uniqueSearch for contacts to actions; add money attributes to track; fix tests

+++ packages/destination-actions/src/destinations/intercom/index.ts
@@ -1,8 +1,9 @@
-import type { DestinationDefinition } from '@segment/actions-core'
+import { DestinationDefinition, IntegrationError } from '@segment/actions-core'
 import type { Settings } from './generated-types'
 import identifyContact from './identifyContact'
 import groupIdentifyContact from './groupIdentifyContact'
 import trackEvent from './trackEvent'
+import { getUniqueIntercomContact } from './util'
 
 const destination: DestinationDefinition<Settings> = {
   name: 'Intercom Cloud Mode (Actions)',
@@ -30,10 +31,20 @@ const destination: DestinationDefinition<Settings> = {
     }
   },
 
+  /**
+   * We search for a unique contact using the userId as the exteralId and then delete. If a contact
+   * is not found, throw a 404
+   */
   onDelete: async (request, { payload }) => {
-    return request(`https://api.intercom.io/contacts/${payload.userId}`, {
-      method: 'DELETE'
-    })
+    const external_id = payload.userId as string
+    const contact = await getUniqueIntercomContact(request, { external_id })
+    if (contact) {
+      return request(`https://api.intercom.io/contacts/${payload.userId}`, {
+        method: 'DELETE'
+      })
+    } else {
+      throw new IntegrationError('No unique contact found', 'Contact not found', 404)
+    }
   },
 
   actions: {
